refactor: モジュールのインライン化
